add jest/supertest to server
