+++ server/index.js
@@ -1,31 +1,18 @@
-const express = require('express');
-const app = express();
 const mongoose = require('mongoose');
-const cors = require('cors');
 
-const mainRoute = require('./routes/index');
-const trafficRoute = require('./routes/traffic');
-const routeRoute = require('./routes/route');
+const app = require('./app');
 
 const config = require('./config.json')
-const host = config.server.host;
-const port = config.server.port;
-const domain = config.domain;
+const { host, port } = config.server;
 
-const corsOptions = {
-  origin: domain
-};
-app.use(cors(corsOptions));app.use(mainRoute);
-
-app.use(trafficRoute);
-app.use(routeRoute);
-
-mongoose.connect('mongodb://localhost:27017/toronto-traffic', {
-  // useFindAndModify: false,
+const DATABASE_NAME = 'toronto-traffic';
+mongoose.connect(`mongodb://localhost:27017/${DATABASE_NAME}`, {
   useNewUrlParser: true,
-  useUnifiedTopology: true
+});
+mongoose.connection.on('connected', () => {
+  console.log(`Connected to '${DATABASE_NAME}' database`)
 });
 
 app.listen(port, () => {
   console.log(`TTCongestion listening at ${host}:${port}`)
-})
+});
