+++ server/package.json
@@ -4,7 +4,8 @@
   "description": "server for Toronto Traffic project",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "jest --config ./jest.config.js",
+    "test:watch": "jest --watch --config ./jest.config.js"
   },
   "author": "",
   "license": "ISC",
@@ -17,6 +18,7 @@
     "cors": "^2.8.5",
     "debug": "^4.3.1",
     "express": "^4.17.1",
+    "jest": "^26.6.3",
     "lodash": "^4.17.20",
     "moment": "^2.29.1",
     "moment-timezone": "^0.5.32",
@@ -26,5 +28,8 @@
     "readline-sync": "^1.4.10",
     "xml2js": "^0.4.23",
     "yargs": "^16.2.0"
+  },
+  "devDependencies": {
+    "supertest": "^6.0.1"
   }
 }
