+++ server/routes/index.js
@@ -1,7 +1,9 @@
 const router = require('express').Router();
 
 router.get('/', (req, res) => {
-    res.send('Hello World!!!')
+  res.send({
+    message: 'Hello World!!!'
+  });
 });
 
 module.exports = router;
